QS_Chatbot/src/qs-chatbot/requirements.txt
pandas==2.2.2
numpy==2.0.1
python-dotenv==1.0.1
pyarrow==17.0.0
streamlit==1.37.0

google-api-core==2.19.1
google-auth==2.32.0
google-cloud==0.34.0
google-cloud-aiplatform==1.59.0
google-cloud-firestore==2.17.0
google-cloud-logging==3.10.0
google-cloud-secret-manager==2.20.1
google-cloud-storage==2.18.0

httpx-oauth==0.3.5
oauth2client==4.1.3


keyring==18.0.1
keyrings.google-artifactregistry-auth==1.0.0
